cometkiwi/requirements.txt
fastapi
uvicorn
pydantic
requests
unbabel-comet>=2.2.0